% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumRates.R
\name{cumRates}
\alias{cumRates}
\title{Extraction of cumulative hit and false alarm rates for memory ROC analysis}
\usage{
cumRates(responseScale, confidenceRatings, TargetLure, targetLabel, lureLabel)
}
\arguments{
\item{responseScale}{An vector containing  possible levels of confidence rating responses ordered from highest to lowest (e.g. 6:1).}

\item{confidenceRatings}{An vector containing participant confidence rating responses to individual items.}

\item{TargetLure}{An vector coding containing information whether each individual item is a Target or Lure item (e.g. Old, New, Old, Old).}

\item{targetLabel}{A string/integer designating the label for the target source in TargetLure.}

\item{lureLabel}{A string/integer designating the label for the lure source in TargetLure.}
}
\value{
The function returns a data frame with components:
\item{fa}{The extracted cumulative false alarm rate.}
\item{hit}{The extracted cumulative hit rate.}
\item{zfa}{The z-transformed cumulative false alarm rate.}
\item{zhit}{The z-transformed cumulative hit rate.}
}
\description{
This function allows you to extract cumulative hit and false alarm rates for memory ROC analysis.
}
\keyword{ROC}
\keyword{alarm}
\keyword{false}
\keyword{hit}
