% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDPSDROC.R
\name{fitDPSDROC}
\alias{fitDPSDROC}
\title{Estimation of recollection and familiarity by fitting recognition memory ROC data with the Dual Process Signal Detection (DPSD) model}
\usage{
fitDPSDROC(falseAlarms, hit, iterations = 200, eqVar = TRUE)
}
\arguments{
\item{falseAlarms}{A vector containing the cumulative recognition false alarm rate.}

\item{hit}{A vector containing the cumulative recognition hit rate.}

\item{iterations}{A numeric value specifying the number of iterations. Default is set to 200.}

\item{eqVar}{A boolean value specifying if the standard deviation of the target distribution is equal to that of the lure distribution (i.e. = 1) (TRUE) or estimated separately (FALSE). Default is set to TRUE.}
}
\value{
The function returns a dataframe with components:
\item{(parameters)}{The estimated parameters (recollection_target, recollection_lure = 0, familiarity, sd_target, criteria) for the iteration with the lowest SumSquareError}
\item{SSE}{Minimum sum square error}
}
\description{
This function allows to estimate recollection and familiarity for recognition memory data by fitting data to the DPSD model.
The optimization is attempted by minimizing the total squared difference between observed and
predicted cumulative hit and false alarm rates using the Broyden-Fletcher-Goldfarb-Shanno (BFGS) algorithm in \code{\link{optim}}.
The function uses random start values on each iteration in order to find the set of parameters,
which fit the data best by returning the values with the lowest total squared difference.
Recollection and Familiarity for the lure distribution (new items) are set to 0.
Optional arguments in the function allow the user to specify an equal-variance model.
Recollection is bounded to be between 0 and 1, Familiarity and the standard deviation of the target distribution to be positive.
Criteria are unbounded.
A high number of iterations is necessary to avoid local minima.
}
\references{
Yonelinas, A. P. (1999). The Contribution of Recollection and Familiarity to Recognition and Source-Memory Judgments: A Formal Dual-Process Model and an Analysis of Receiver Operating Characteristics. Journal of Experimental Psychology: Learning, Memory, and Cognition, 25(6), 1415 - 1434. http://doi.org/10.1037//0278-7393.25.6.1415
}
\author{
Nicholas Lange, \email{lange.nk@gmail.com}
}
\keyword{DPSD}
\keyword{ROC}
\keyword{familiarity}
\keyword{recollection}
