% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMLEstartparameters.R
\name{makeMLEstartparameters}
\alias{makeMLEstartparameters}
\title{Dual process MLE: Make start parameters}
\usage{
makeMLEstartparameters(freqFA, freqHit, iterations)
}
\arguments{
\item{freqFA}{frequency of FA respones in category bins, Lure to Target order}

\item{freqHit}{frequency of Hit responses in category bins, Lure to Target order}

\item{iterations}{the number of runs}
}
\value{
The function returns a dataframe:
\item{(parameters)}{Starting parameters for all parameters that versions of the DP Recog-Source model could vary to date}
}
\description{
This function estimates start parameters for the DP univariate model fit, with starting parameters for the
first 40% of iterations sampled from around the observed data, and 60% sampled from a uniform distribution.
}
\author{
Nicholas Lange, \email{lange.nk@gmail.com}
}
\keyword{dual}
\keyword{likelihood}
\keyword{maximum}
\keyword{process,}
