% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictedROC.R
\name{predictedROC}
\alias{predictedROC}
\title{Returning false alarm and hit rates for the Dual Process Signal Detection (DPSD) for recognition and source memory}
\usage{
predictedROC(recollection_target, recollection_lure, familiarity,
  sd_target = 1, fitSource = TRUE)
}
\arguments{
\item{recollection_target}{A value representing the recollection rate for target items.}

\item{recollection_lure}{A value representing the recollection rate for lure items.}

\item{familiarity}{A value representing the familiarity rate}

\item{sd_target}{A value representing the standard deviation of the target distribution (default: sd_target = 1)}

\item{fitSource}{A boolean value representing whether a recognition or source memory ROC is fitted (default: fitSource = TRUE)}
}
\value{
The function returns a data frame with components:
\item{fa}{The fitted cumulative false alarm rate.}
\item{hit}{The fitted cumulative hit rate.}
}
\description{
This function allows you to get the correspending false alarm and hit rates for a given set of parameters (recollection
for both distributions, familiarity, standard deviation for the target distribution). A boolean
allows you to fit a source ROC when set TRUE, with target and lure distributions representing designated target source
and lure source, and a recognition ROC when set FALSE (with recollection_lure and familiarity set to 0 for the lure
distribution).
}
\references{
Yonelinas, A. P. (1999). The Contribution of Recollection and Familiarity to Recognition and Source-Memory Judgments: A Formal Dual-Process Model and an Analysis of Receiver Operating Characteristics. Journal of Experimental Psychology: Learning, Memory, and Cognition, 25(6), 1415 - 1434. http://doi.org/10.1037//0278-7393.25.6.1415
}
\author{
Nicholas Lange, \email{lange.nk@gmail.com}
}
\keyword{ROC}
\keyword{memory}
