% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratingFreg.R
\name{ratingFreq}
\alias{ratingFreq}
\title{Extraction of cumulative hit and false alarm rates for memory ROC analysis}
\usage{
ratingFreq(responseScale, confidenceRatings, TargetLure, targetLabel, lureLabel)
}
\arguments{
\item{responseScale}{An vector containing  possible levels of confidence rating responses ordered from highest to lowest (e.g. 6:1).}

\item{confidenceRatings}{An vector containing participant confidence rating responses to individual items.}

\item{TargetLure}{An vector coding containing information whether each individual item is a Target or Lure item (e.g. Old, New, Old, Old).}

\item{targetLabel}{A string/integer designating the label for the target source in TargetLure.}

\item{lureLabel}{A string/integer designating the label for the lure source in TargetLure.}
}
\value{
The function returns a data frame with components:
\item{fa}{The extracted frequency of hit responses per confidence rating.}
\item{hit}{The extracted frequency of fa responses per confidence rating.}
}
\description{
This function allows you to extract cumulative hit and false alarm rates for memory ROC analysis.
}
